refactor spec2cases parsing and validation modules
